server/requirements.txt
tensorflow==1.6.0
requests==2.20.0
python-firebase==1.2.0